+++ src/pages/PostsManagerPage.tsx
@@ -2,8 +2,9 @@ import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from
 import { useEffect, useState } from "react"
 import { useLocation, useNavigate } from "react-router-dom"
 
+import { Button } from "@shared/ui"
+
 import {
-  Button,
   Card,
   CardContent,
   CardHeader,
